refactor: item 패키지 분리
