+++ back/src/test/java/com/gagi/market/item/service/ItemServiceTest.java
@@ -1,8 +1,8 @@
-package com.gagi.market.service;
+package com.gagi.market.item.service;
 
-import com.gagi.market.api.dto.ItemRequestDto;
-import com.gagi.market.domain.Item;
-import com.gagi.market.domain.ItemRepository;
+import com.gagi.market.item.api.dto.ItemRequestDto;
+import com.gagi.market.item.domain.Item;
+import com.gagi.market.item.domain.ItemRepository;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
